added correct status code for every endpoint

+++ src/controller/attendance/attendanceDelete.js
@@ -28,9 +28,15 @@ const deleteRecord = async (req, res) => {
 const deleteAll = async(req,res)=>{
     try {
         
+        const attendance = await Attendance.find();
+
+        if(!attendance){
+            return res.status(404).send("No record found!");
+        }
+        
         await Attendance.deleteMany();
 
-        res.status(200).send("All the attendance have been deleted!");
+        res.status(204).send("All the attendance have been deleted!");
 
     } catch (error) {
